+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.108.0] - Not released
+### Added
+- Links to the user/group memories in the sidebar of the user/group-related
+  pages
+- Click on the ^^^-reference shows a preview of the referenced comment
+
+### Fixed
+- Remove usage of "screen name" in favor of "display name" in UI
+
+### Changed
+- Improve texts for account privacy settings
+
 ## [1.107.1] - 2022-03-24
 ### Changed
 - Show registration date in Profiles and Groups